RxJS/ContextHook [Small Refactor]
