+++ src/reactivity/common/operations.ts
@@ -1,15 +1,16 @@
 export function incrementDecrement(operation: string, value: number) {
-  let newValue = 0
-
   switch (operation) {
     case 'increment':
-      newValue = value + 1
+      value++
       break
 
     case 'decrement':
-      newValue = value - 1
+      value--
       break
+
+    default:
+      value = 0
   }
 
-  return newValue
+  return value
 }
